feat(i18n): restructure app directory under [locale] segment
