+++ src/app/[locale]/(app)/admin/page.tsx
