+++ src/app/[locale]/(app)/dashboard/page.tsx
