+++ src/app/[locale]/(app)/groups/[groupId]/page.tsx
