+++ src/app/[locale]/(app)/groups/[groupId]/scan/page.tsx
