+++ src/app/[locale]/(app)/layout.tsx
