+++ src/app/[locale]/(app)/settings/page.tsx
