+++ src/app/[locale]/(auth)/verify-request/page.tsx
