+++ src/app/[locale]/invite/[token]/page.tsx
