+++ src/app/[locale]/split/[token]/page.tsx
