+++ src/app/[locale]/split/layout.tsx
