+++ src/app/[locale]/split/page.tsx
